fix hint issue
