+++ scripts/controlnet_ui/tool_button.py
@@ -1,0 +1,12 @@
+import gradio as gr
+
+class ToolButton(gr.Button, gr.components.FormComponent):
+    """Small button with single emoji as text, fits inside gradio forms"""
+
+    def __init__(self, **kwargs):
+        super().__init__(variant="tool", 
+                         elem_classes=kwargs.pop('elem_classes', []) + ["cnet-toolbutton"], 
+                         **kwargs)
+
+    def get_block_name(self):
+        return "button"
